feat: move FluxClient to flux/client.py with timeout and context manager support
Extract FluxClient from flux/console/client.py to flux/client.py as a
top-level module. Add configurable timeout parameter (default 10s) and
async context manager protocol (__aenter__/__aexit__). The console module
re-exports FluxClient for backward compatibility.
